// src/app/api/models/Warehouses/warehouse-response.ts
export interface WarehouseResponse {
    id: string;
    name: string;
    address: string;
    phone: string;
    warehouseType: string;
    number: number;
    cityDescription: string;
    settlementAreaDescription: string;
    settlementRegionsDescription: string;
    settlementTypeDescription: string;
    warehouseCategory: string;
    settlementId: string;
    deliveryId: string;
}